(refactor) : 게임을 진행하는 책임을 Round에게 이동
