+++ src/main/java/racingcar/domain/Cars.java
@@ -20,7 +20,7 @@ public class Cars {
     }
 
     public static Cars of(String inputCarNames) {
-        List<Name> carNames = Arrays.stream(inputCarNames.split(",")).map(Name::valueOf).collect(Collectors.toList());
+        List<Name> carNames = Arrays.stream(inputCarNames.split(DELIMITER)).map(Name::valueOf).collect(Collectors.toList());
         List<Car> cars = carNames.stream().map(Car::of).collect(Collectors.toList());
         return new Cars(cars);
     }
